`GamesRest` extracted to common repository
