+++ app/src/main/java/edu/agh/susgame/dto/rest/games/model/Request.kt
@@ -3,12 +3,11 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-# Current API Version
-API version: `0.2`
+package edu.agh.susgame.dto.rest.games.model
 
 
-# Changelog
-
-* ...
-* `0.2` - REST models extracted to common repository
-* `0.1` - Api was initialized
+data class GameCreationRequest(
+    val gameName: String,
+    val maxNumberOfPlayers: Int = 4,
+    val gamePin: String? = null,
+)
