+++ app/src/main/java/edu/agh/susgame/dto/rest/games/model/Result.kt
@@ -1,4 +1,9 @@
-package edu.agh.susgame.front.service.web.rest.games.model
+// WARNING: THIS FILE WAS CLONED AUTOMATICALLY FROM 'SusGameDTO' GITHUB REPOSITORY
+// IT SHOULD NOT BE EDITED IN ANY WAY
+// IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
+// IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
+
+package edu.agh.susgame.dto.rest.games.model
 
 import edu.agh.susgame.dto.rest.model.Lobby
 import edu.agh.susgame.dto.rest.model.LobbyId
