+++ app/src/main/java/edu/agh/susgame/front/service/web/WebLobbyService.kt
@@ -1,17 +1,17 @@
 package edu.agh.susgame.front.service.web
 
+import edu.agh.susgame.dto.rest.games.GamesRest
+import edu.agh.susgame.dto.rest.games.model.CreateGameApiResult
+import edu.agh.susgame.dto.rest.games.model.GetAllGamesApiResult
+import edu.agh.susgame.dto.rest.games.model.GetGameApiResult
 import edu.agh.susgame.dto.rest.model.Lobby
 import edu.agh.susgame.dto.rest.model.LobbyId
 import edu.agh.susgame.front.service.interfaces.CreateNewGameResult
 import edu.agh.susgame.front.service.interfaces.LobbyService
-import edu.agh.susgame.front.service.web.rest.games.GamesRestImpl
-import edu.agh.susgame.front.service.web.rest.games.model.CreateGameApiResult
-import edu.agh.susgame.front.service.web.rest.games.model.GetAllGamesApiResult
-import edu.agh.susgame.front.service.web.rest.games.model.GetGameApiResult
 import java.util.concurrent.CompletableFuture
 
 
-class WebLobbyService(private val gamesRest: GamesRestImpl) : LobbyService {
+class WebLobbyService(private val gamesRest: GamesRest) : LobbyService {
     override fun getAll(): CompletableFuture<Map<LobbyId, Lobby>> =
         gamesRest.getAllGames().thenApply { response ->
             when (response) {
