--- app/src/main/java/edu/agh/susgame/front/service/web/rest/games/model/Response.kt
@@ -1,6 +1,0 @@
-package edu.agh.susgame.front.service.web.rest.games.model
-
-
-data class GameCreationApiResponse(
-    val gameId: Int,
-)
